Fix users from being logged out when hitting login page
Two things were causing this: If users were logged in, the would be
redirected, but this activated on mount when login had not been
confirmed by the cookies yet. If they weren't logged in (which was
assumed by the logic, the cookies were deleted, thus actually logging
them out).

Additionally, the logic there would not re-activate when the user was
verified to be logged in, if ever (it only ran once on component mount).

Lastly, there was a onMount useEffect that cleared the cookies anyways.

I don't know how this wasn't caught yet. But now it's fixed.
There is still a moment where the login page is shown regardless of
login, but that's fine for now. TODO is marked for this.