Add column visibility system to EventsView, VisibilityDropdown component
This was quite an involved component as I wasn't too familiar with
Headless UI or Tanstack's respective components, so I had to spend a lot
of time fiddling with both of them in order to connect them together.

I think the solution I've come up with looks nice and is decently
optimized for now.