Add documentation to Breadcrumbs component, allow ReactNode labels, use index for loop keys
I have to use the index as the loop key, as ReactNode allows nullish values, and with the label as the key,
this doesn't work for the index, which can't use nullish values. So we switch the key to the index.
Unsure how I feel about using JSDoc for comments on the purpose of the Breadcrumb properties...