NoSSR comments, fix double usage of useGlobalContext(), fix formatString
Unsure why, but lightFormat was accepting a six-letter 'a' specifier for the date format string.
This shouldn't be valid, but it matched the four-letter 'a' specifier variant, so I put that instead.
NoSSR seems a little weird and difficult to digest, so I think more comments & documentation are necessary.
I hope I was right about how NoSSR & toLocaleString are used here ;P