Slightly separate event buttons, remove not-logged-in view on check-in button
So I added some text for when a user isn't logged in but check-in is available,
but I'm scrapping it for now because it requires more logic to be bullet proof.
Easier to not add authentication as another dimension to the view of a single button.
Also, I separated the event's buttons because this page is huge.